+++ tests/types.go
@@ -1,0 +1,24 @@
+package tests
+
+import (
+	"k8s.io/apimachinery/pkg/runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+type SolaceScalableReconciler struct {
+	client.Client
+	Scheme *runtime.Scheme
+}
+
+/*
+type k8sClient interface {
+	Get(
+		ctx context.Context,
+		key types.NamespacedName,
+		obj client.Object,
+		opts ...client.GetOption,
+	) error
+	Create(ctx context.Context, obj client.Object, opts ...client.CreateOption) error
+	Update(ctx context.Context, obj client.Object, opts ...client.UpdateOption) error
+}
+*/
